Fixes performance issue.
On Chrome opening and closing the top expansion panels was a little slow.
On Firefox it was absolutely horrible. Every time the expansion panels
opened, the board was pushed down. As a result the board needed to be
repainted, and it was that painting which was causing the slow down.
It turns out that the drop-shadow filter on the SVG was the culprit.
It turns out drop shadows are incredibly computationally expensive,
which became an issue on such a large intricate shape. Once the
drop-shadow was removed the problem was solved.